requirements.txt
streamlit
pandas
plotly.express
json
plotly.graph_objects
geopandas
